Replace hardcoded chown user with USER variable
I tried running this on a machine of mine and I got an error because there was no `bennett` user on my host machine. Changing this to use `$USER` made the docker build step work without erroring out and applied the correct permissions to the privacybadger directory so that the application could write to it in the docker container and I could view it on the host machine.