chown pbpath in RUN command instead of COPY
On docker 18, the $USER argument to COPY --chown=$USER:$USER wasn't being
evaluated as a shell variable, so it was failing to find the user "\$USER".

By running the chown afterwards in the container, the directory is chowned to
the user specified in the Dockerfile. In order for the chown to run properly
we have to change the USER back to root (since the files are owned by root after
the copy) and then change the USER back to the UNAME value afterwards.